+++ src/rdbus/config/wago/Module.hpp
@@ -5,7 +5,7 @@
 #include <nlohmann/json.hpp>
 #include <string>
 
-namespace rdbus::config::ip
+namespace rdbus::config::wago
 {
 
 struct Module
@@ -18,9 +18,11 @@ struct Module
     std::list< std::string > instances;
 
     // Absolute instance offset from all fields of other instances
-    unsigned int offset = 0;
+    // Instances start at 1
+    constexpr static unsigned int defaultOffset = 1;
+    unsigned int offset = defaultOffset;
 };
 
 void from_json( const nlohmann::json& j, Module& x );
 
-} // namespace rdbus::config::ip
+} // namespace rdbus::config::wago