+++ tests/serializers/eip/LimitsTests.cpp
@@ -1,13 +1,13 @@
 #include "rdbus/config/Exception.hpp"
-#include "rdbus/config/ip/Limits.hpp"
+#include "rdbus/config/wago/Limits.hpp"
 #include "tests/utility.hpp"
 #include <gtest/gtest.h>
 #include <nlohmann/json.hpp>
 #include <testDataDirectory.hpp>
 
-const std::string testFilePath = TEST_DATA_DIR "/serializers/json_files/ip/limits/";
+const std::string testFilePath = TEST_DATA_DIR "/serializers/json_files/wago/limits/";
 using namespace nlohmann;
-using namespace rdbus::config::ip;
+using namespace rdbus::config::wago;
 using namespace rdbus;
 
 TEST( TestLimits, Valid )